prevent corruption in Tags.fromValue on Scala.js
