+++ core/jvm/src/main/scala/zio/ChunkPlatformSpecific.scala
@@ -1,0 +1,56 @@
+package zio
+
+import scala.reflect.{ classTag, ClassTag }
+
+private[zio] trait ChunkPlatformSpecific {
+
+  private[zio] object Tags {
+    def fromValue[A](a: A): ClassTag[A] =
+      unbox(ClassTag(a.getClass))
+
+    private def unbox[A](c: ClassTag[A]): ClassTag[A] =
+      if (isBoolean(c)) BooleanClass.asInstanceOf[ClassTag[A]]
+      else if (isByte(c)) ByteClass.asInstanceOf[ClassTag[A]]
+      else if (isShort(c)) ShortClass.asInstanceOf[ClassTag[A]]
+      else if (isInt(c)) IntClass.asInstanceOf[ClassTag[A]]
+      else if (isLong(c)) LongClass.asInstanceOf[ClassTag[A]]
+      else if (isFloat(c)) FloatClass.asInstanceOf[ClassTag[A]]
+      else if (isDouble(c)) DoubleClass.asInstanceOf[ClassTag[A]]
+      else if (isChar(c)) CharClass.asInstanceOf[ClassTag[A]]
+      else classTag[AnyRef].asInstanceOf[ClassTag[A]] // TODO: Find a better way
+
+    private def isBoolean(c: ClassTag[_]): Boolean =
+      c == BooleanClass || c == BooleanClassBox
+    private def isByte(c: ClassTag[_]): Boolean =
+      c == ByteClass || c == ByteClassBox
+    private def isShort(c: ClassTag[_]): Boolean =
+      c == ShortClass || c == ShortClassBox
+    private def isInt(c: ClassTag[_]): Boolean =
+      c == IntClass || c == IntClassBox
+    private def isLong(c: ClassTag[_]): Boolean =
+      c == LongClass || c == LongClassBox
+    private def isFloat(c: ClassTag[_]): Boolean =
+      c == FloatClass || c == FloatClassBox
+    private def isDouble(c: ClassTag[_]): Boolean =
+      c == DoubleClass || c == DoubleClassBox
+    private def isChar(c: ClassTag[_]): Boolean =
+      c == CharClass || c == CharClassBox
+
+    private val BooleanClass    = classTag[Boolean]
+    private val BooleanClassBox = classTag[java.lang.Boolean]
+    private val ByteClass       = classTag[Byte]
+    private val ByteClassBox    = classTag[java.lang.Byte]
+    private val ShortClass      = classTag[Short]
+    private val ShortClassBox   = classTag[java.lang.Short]
+    private val IntClass        = classTag[Int]
+    private val IntClassBox     = classTag[java.lang.Integer]
+    private val LongClass       = classTag[Long]
+    private val LongClassBox    = classTag[java.lang.Long]
+    private val FloatClass      = classTag[Float]
+    private val FloatClassBox   = classTag[java.lang.Float]
+    private val DoubleClass     = classTag[Double]
+    private val DoubleClassBox  = classTag[java.lang.Double]
+    private val CharClass       = classTag[Char]
+    private val CharClassBox    = classTag[java.lang.Character]
+  }
+}
